Revert "Change import structure for editable installs."
While executing script from subprocess, this creates import errors as
the package maynot be installed. Since we intend to start server from
other language than python, this won't be used for now.

Importing directly wihtout `.` will cause problems for imports used in
tests. For that, editable install is requied.

But the recommended way is to install the server as separate package and
decopule the workings from vscode as to make it integrateable into other
IDEs in the future (if required).

This reverts commit a86388ce23139641f8f687e5fbfa2f31180218a3.